mysite/requirements.txt
Django==3.2.14
djangorestframework==3.13.1
coverage==6.4.1
celery==5.1.2
